+++ changelog.md
@@ -1,2 +1,2 @@
 ### Simple Magnets 1.1.12
-- Initial release of Simple Magnets for Minecraft 1.21.9 & 1.21.10
+- Initial release of Simple Magnets for Minecraft 1.21.11